+++ .docs/changelog.md
@@ -7,6 +7,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED](https://gitlab.phaidra.org/fair-data-austria-db-repository/fda-services/-/tags/v1.9.2) - 2025-06-??
+
+#### Features
+
+* Added an endpoint to check for database grants and display them in the UI (Database > Settings) for the owner 
+  in [#536](https://gitlab.phaidra.org/fair-data-austria-db-repository/fda-services/-/issues/536).
+
 ## [v1.9.1](https://gitlab.phaidra.org/fair-data-austria-db-repository/fda-services/-/tags/v1.9.1) - 2025-05-31
 
 #### Changes